examples/diffusers/sdxl/requirements_sdxl_extra.txt
opencv-python
scipy==1.10.1
torchvision==0.16.0
toml
voluptuous
deepspeed
decorator==4.4.2
pybind11
ninja
wheel
numpy
six
regex
attrs
psutil
pyyaml
protobuf
einops
sentencepiece
pytest
gpytorch
pandas